backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-environ==0.11.2
django-cors-headers==4.3.1
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
python-dateutil==2.8.2
django-filter==23.5
argon2-cffi==23.1.0
django-redis==5.4.0
whitenoise==6.6.0
